+++ test/Sentry/EventHandler/DatabaseEventsTest.php
@@ -4,6 +4,8 @@ namespace Sentry\Laravel\Tests\EventHandler;
 
 use Illuminate\Database\Connection;
 use Illuminate\Database\Events\QueryExecuted;
+use Illuminate\Database\Grammar;
+use Illuminate\Database\Query\Builder;
 use Mockery;
 use Sentry\Laravel\Tests\TestCase;
 
@@ -46,11 +48,56 @@ class DatabaseEventsTest extends TestCase
 
         $lastBreadcrumb = $this->getLastSentryBreadcrumb();
 
-        $this->assertEquals($query, $lastBreadcrumb->getMessage());
-        $this->assertEquals($bindings, $lastBreadcrumb->getMetadata()['bindings']);
+        $this->assertSame($query, $lastBreadcrumb->getMessage());
+        $this->assertSame($bindings, $lastBreadcrumb->getMetadata()['bindings']);
     }
 
-    public function testSqlQueriesAreRecordedWhenDisabled(): void
+    public function testSqlBindingsAreInlined(): void
+    {
+        $this->resetApplicationWithConfig([
+            'sentry.breadcrumbs.sql_bindings' => 'embed',
+        ]);
+
+        $this->assertSame('embed', $this->app['config']->get('sentry.breadcrumbs.sql_bindings'));
+
+        $query = 'SELECT * FROM breadcrumbs WHERE bindings = ?;';
+        $bindings = ['1'];
+        $rawSQL = 'SELECT * FROM breadcrumbs WHERE bindings = 1;';
+
+        $grammar = Mockery::mock(Grammar::class);
+        $grammar
+            ->shouldReceive('substituteBindingsIntoRawSql')
+            ->with($query, $bindings)
+            ->andReturn($rawSQL);
+
+        $builder = Mockery::mock(Builder::class);
+        $builder
+            ->shouldReceive('getGrammar')
+            ->andReturn($grammar);
+
+        $connection = Mockery::mock(Connection::class);
+        $connection->shouldReceive('getName')
+            ->andReturn('test');
+        $connection->shouldReceive('query')
+            ->andReturn($builder);
+        $connection->shouldReceive('prepareBindings')
+            ->with($bindings)
+            ->andReturn($bindings);
+
+        $this->dispatchLaravelEvent(new QueryExecuted(
+            $query,
+            $bindings,
+            10,
+            $connection
+        ));
+
+        $lastBreadcrumb = $this->getLastSentryBreadcrumb();
+
+        $this->assertSame($rawSQL, $lastBreadcrumb->getMessage());
+        $this->assertArrayNotHasKey('bindings', $lastBreadcrumb->getMetadata());
+    }
+
+    public function testSqlQueriesAreNotRecordedWhenDisabled(): void
     {
         $this->resetApplicationWithConfig([
             'sentry.breadcrumbs.sql_queries' => false,
@@ -68,7 +115,7 @@ class DatabaseEventsTest extends TestCase
         $this->assertEmpty($this->getCurrentSentryBreadcrumbs());
     }
 
-    public function testSqlBindingsAreRecordedWhenDisabled(): void
+    public function testSqlBindingsAreNotRecordedWhenDisabled(): void
     {
         $this->resetApplicationWithConfig([
             'sentry.breadcrumbs.sql_bindings' => false,
@@ -85,13 +132,14 @@ class DatabaseEventsTest extends TestCase
 
         $lastBreadcrumb = $this->getLastSentryBreadcrumb();
 
-        $this->assertEquals($query, $lastBreadcrumb->getMessage());
-        $this->assertFalse(isset($lastBreadcrumb->getMetadata()['bindings']));
+        $this->assertSame($query, $lastBreadcrumb->getMessage());
+        $this->assertArrayNotHasKey('bindings', $lastBreadcrumb->getMetadata());
     }
 
     private function getMockedConnection()
     {
         return Mockery::mock(Connection::class)
-            ->shouldReceive('getName')->andReturn('test');
+            ->shouldReceive('getName')
+            ->andReturn('test');
     }
 }